fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
cryptography==41.0.7
httpx==0.25.2
email-validator==2.1.0
